factor out claim extraction into its own module
